pyproject.toml
[project]
name = "learn-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.5.0",
    "google-generativeai>=0.8.5",
    "litellm>=1.73.2",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.1",
    "yfinance>=0.2.64",
]
